+++ packages/cli/src/commands/index.ts
@@ -1,8 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './createContentBlock.js';
-export * from './createPlatformApp.js';
 export * from './deploy.js';
 export * from './login.js';
 export * from './logout.js';
 export * from './serve.js';
+export * from './create.js';
